+++ src/TEMPO.jl
@@ -37,7 +37,7 @@ export rdm, quantummap, meanforcestate, mfs
 # influence functional
 export HybridizationStyle, AdditiveHyb, NonAdditiveHyb, NonDiagonalHyb, pairop
 export PartialIF, XTRGIF, TDVPIF
-export influenceoperator, influenceoperatorexponential, differentialinfluencefunctional
+export influenceoperators, influenceoperatorsteppers, influenceoperatorstepper
 export hybriddynamics, hybriddynamics!, hybriddynamics_naive, hybriddynamics_naive!
 export partialif_naive, partialif
 # boundary condition