+++ src/influencefunctional/ttiif/pt/pt.jl
@@ -41,10 +41,10 @@ end
 function _hybriddynamics_fast(lattice::AbstractPTLattice, corr::AbstractCorrelationFunction, hyb::GeneralHybStyle, alg::XTRGIF)
 	algmult = alg.algmult
 	if alg.verbosity > 1
-		t = @elapsed mps = differentialinfluencefunctional(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algmult, algexpan=alg.algexpan)
+		t = @elapsed mps = influenceoperatorstepper(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algmult, algexpan=alg.algexpan)
 		println("building the initial MPS-IF takes $t seconds, bond dimension is ", bond_dimension(mps))
 	else
-		mps = differentialinfluencefunctional(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algmult, algexpan=alg.algexpan)
+		mps = influenceoperatorstepper(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algmult, algexpan=alg.algexpan)
 	end
 	
 	for i in 1:alg.k
@@ -61,10 +61,10 @@ end
 function _hybriddynamics_slow(lattice::AbstractPTLattice, corr::AbstractCorrelationFunction, hyb::GeneralHybStyle, alg::XTRGIF)
 	algmult = alg.algmult
 	if alg.verbosity > 1
-		t = @elapsed mps0 = differentialinfluencefunctional(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algmult, algexpan=alg.algexpan)
+		t = @elapsed mps0 = influenceoperatorstepper(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algmult, algexpan=alg.algexpan)
 		println("building the initial MPS-IF takes $t seconds, bond dimension is ", bond_dimension(mps))
 	else
-		mps0 = differentialinfluencefunctional(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algmult, algexpan=alg.algexpan)
+		mps0 = influenceoperatorstepper(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algmult, algexpan=alg.algexpan)
 	end
 	mps = mps0
 
@@ -82,10 +82,10 @@ end
 function _hybriddynamics_slow!(gmps, lattice::AbstractPTLattice, corr::AbstractCorrelationFunction, hyb::GeneralHybStyle, alg::XTRGIF)
 	algmult = alg.algmult
 	if alg.verbosity > 1
-		t = @elapsed mps_all = influenceoperatorexponential(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algexpan=alg.algexpan)
+		t = @elapsed mps_all = influenceoperatorsteppers(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algexpan=alg.algexpan)
 		println("building the initial MPS-IFs takes $t seconds")
 	else
-		mps_all = influenceoperatorexponential(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algexpan=alg.algexpan)
+		mps_all = influenceoperatorsteppers(lattice, corr, 1/2^(alg.k), hyb, alg.algevo, algexpan=alg.algexpan)
 	end
 
 	for i in 1:2^(alg.k)